src/requirements.txt
cloudshell-networking-devices>=2.3,<2.4
cloudshell-firewall>=3.0,<3.1
